% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abtest1}
\alias{abtest1}
\title{abtest1}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{countryID}{ID of visitor's country, factor}
  \item{latitude}{last position of visitor, recorded before the test, numeric}
  \item{longitude}{last position of visitor, recorded before the test, numeric}
  \item{langID}{ID of visitor language, factor}
  \item{name}{visitor's OS, factor}
  \item{device}{ID of visitor device, factor}
  \item{userAgent}{visitor's user agent, factor}
  \item{A}{visitor's reward with A version of a website}
  \item{B}{visitor's reward with A version of a website}
  ...
}}
\source{
\url{http://abtasty.com/}
}
\usage{
abtest1
}
\description{
A/B test on 6216 visitors of a website (Learnset:30\% Trainset 70\%).
Dataset from a frequentist ab test conducted in 2018 by a merchant website.
Visitors to the site have seen version A or version B only.
 30% of the dataset is used for a pre-processing step (see CTREEUCB).
 The data is not modified (missing values).
 set learn_size = 1865
 The remaining 70% of the dataset is dedicated to the online phase.
 It is necessary to have a reward for all variations.
 Missing values are replaced by bootstrap with the Leave at Least One method
}
\examples{
 try(data(package = "bandit4abtest") )
 data(abtest1)
}
\keyword{datasets}
