% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageRegret.R
\name{averageRegret}
\alias{averageRegret}
\title{Return list of regret}
\usage{
averageRegret(
  choice,
  visitor_reward,
  dt,
  ctree_models,
  isRewardAreBoolean = FALSE
)
}
\arguments{
\item{choice}{Integer list}

\item{visitor_reward}{dataframe of integer or numeric values}
}
\value{
List of numeric values
}
\description{
Return a list with obtained regret at each iterations
}
\examples{
##### Pairewise #####
set.seed(1234)
size.tot <- 10000
x <- seq(0, 5, 0.01)
x1<- sample(x, size.tot, replace = TRUE, prob = NULL)
arm_1 <-  as.vector(c(2,-1,1.5,0))
K1 <- (x1 < 1 ) * arm_1[4]  +
 (x1 >= 1 & x1 < 2 ) * arm_1[1]  +
 (x1 >= 2 & x1 < 3) * arm_1[2]  +
 (x1 >= 3 & x1 < 4) * arm_1[3]  +
 (x1 >= 4) * arm_1[4]
plot(x1, K1)

arm_2 <-  as.vector(c(1.5,-0.5,1.25,0))
K2 <- (x1 < 1 ) * arm_2[4]  +
 (x1 >= 1 & x1 < 2 ) * arm_2[1]  +
 (x1 >= 2 & x1 < 3) * arm_2[2]  +
 (x1 >= 3 & x1 < 4) * arm_2[3]  +
 (x1 >= 4) * arm_2[4]
plot(x1, K2)
#covariate without interest
x2<- sample(x, size.tot, replace = TRUE, prob = NULL)
#Results for each variation
visitor_reward <-  data.frame(K1,K2 )
summary(visitor_reward)
dt <- as.data.frame(cbind(x1,x2))
#Random choices
choice <- sample(c(1,2), size.tot, replace = TRUE)

#Best policy
ctree_models <- c()
 for(i in 1:ncol(visitor_reward)){
### learning  ###
#Generate formula and tree
ctree_models[[i]] <-  ctree_formula_generate(dt = dt,
                                            visitor_reward = visitor_reward,
                                            ctree_control_val = ctree_control(teststat = c("quadratic")),
                                            arm_for_learn = colnames(visitor_reward[i]),
                                            explanatory_variable= c("x1","x2"),
                                            learn_size = nrow(dt))
}
regret <- averageRegret(choice=choice,visitor_reward=visitor_reward,dt,ctree_models)
plot(1:size.tot,  cumsum(regret))
}
