% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit_reward_control.R
\name{bandit_reward_control}
\alias{bandit_reward_control}
\title{bandit_reward_control}
\usage{
bandit_reward_control(visitorReward, K = ncol(visitorReward))
}
\arguments{
\item{visitorReward}{Dataframe of integer or numeric values}

\item{K}{Integer value (optional)}
}
\value{
Logical value
}
\description{
Control data for bandit algorithm.
See also \code{\link{control_data_missing}} and \code{\link{data_control_K}}
}
\examples{
## Generates 1000 numbers from 2 uniform distributions
set.seed(4434)
K1 <- rnorm(100, 30, .05)
K2 <- rnorm(100, 21, .05)
## Define a dataframe of rewards
visitorReward <- as.data.frame(cbind(K1,K2) )
## Control
bandit_reward_control(visitorReward,K=3)
## Control
bandit_reward_control(visitorReward,K=2)
visitorReward[1,1]= NA
## Control
bandit_reward_control(visitorReward)

}
