% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit_reward_control.R
\name{BanditRewardControl}
\alias{BanditRewardControl}
\title{BanditRewardControl}
\usage{
BanditRewardControl(visitor_reward, K = ncol(visitor_reward))
}
\arguments{
\item{visitor_reward}{Dataframe of integer or numeric values}

\item{K}{Integer value (optional)}
}
\value{
Logical value
}
\description{
Control data for bandit algorithm.
See also \code{\link{ControlDataMissing}} and \code{\link{DataControlK}}
}
\examples{
## Generates 1000 numbers from 2 uniform distributions
set.seed(4434)
K1 <- rnorm(100, 30, .05)
K2 <- rnorm(100, 21, .05)
## Define a dataframe of rewards
visitor_reward <- as.data.frame(cbind(K1,K2) )
## Control
BanditRewardControl(visitor_reward,K=2)

}
