% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_data_type_for_ctree.R
\name{change_data_type_for_ctree}
\alias{change_data_type_for_ctree}
\title{change_data_type_for_ctree}
\usage{
change_data_type_for_ctree(dt, is_reward_are_boolean = FALSE,
  visitorReward)
}
\arguments{
\item{dt}{Dataframe of integer numeric or factor values}

\item{is_reward_are_boolean}{logical value (optional)}

\item{visitorReward}{Dataframe of integer or numeric values}
}
\value{
List with the updated dataframe dt and the updated visitorReward
}
\description{
Check if a reward is defined as logical value: change reward type as factor.
Check if a a colonm of covariates is caractere or logical: change type as factor.
}
\examples{
## Generates 1000 numbers from 2 uniform distributions
c1 <- rep("Test",100)
dt <- as.data.frame(c1)
typeof(change_data_type_for_ctree(dt=dt,visitorReward=dt)$visitorReward[,1])
K1 <- sample(c(0,1),replace=TRUE,size= 100)
visitorReward <- as.data.frame(K1)
typeof(change_data_type_for_ctree(dt=dt,visitorReward=dt,is_reward_are_boolean=FALSE)$visitorReward[,1])
}
