% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeDataTypeForCtreeUCB.R
\name{changeDataTypeForCtreeUCB}
\alias{changeDataTypeForCtreeUCB}
\title{changeDataTypeForCtreeUCB}
\usage{
changeDataTypeForCtreeUCB(dt, is_reward_are_boolean = FALSE, visitor_reward)
}
\arguments{
\item{dt}{Dataframe of integer numeric or factor values}

\item{is_reward_are_boolean}{logical value (optional)}

\item{visitor_reward}{Dataframe of integer or numeric values}
}
\value{
List with the updated dataframe dt and the updated visitor_reward
}
\description{
Check if a reward is defined as logical value: change reward type as factor.
Check if a a colonm of covariates is caractere or logical: change type as factor.
}
\examples{
c1 <- rep("Test",100)
dt <- as.data.frame(c1)
K1 <- sample(c(0,1),replace=TRUE,size= 100)
visitor_reward <- as.data.frame(K1)
temp <- changeDataTypeForCtreeUCB(dt=dt,visitor_reward=visitor_reward,is_reward_are_boolean=TRUE)
is.factor(temp$dt$c1)
is.factor(temp$visitor_reward[,1])
}
