% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_type.R
\name{ChangeType}
\alias{ChangeType}
\title{Change all colomns of a dataframe into
numerical type}
\usage{
ChangeType(visitor_reward)
}
\arguments{
\item{visitor_reward}{a dataframe}
}
\value{
dataframe of numerical value
}
\description{
Change all colomns of a dataframe into
numerical type
}
\examples{
## Generates 1000 numbers from 2 binomial distributions
set.seed(4434)
K1 <- rbinom(1000, 1, 0.6)
K2 <- rbinom(1000, 1, 0.7)
## Define a dataframe of rewards
visitor_reward <- as.data.frame( cbind(as.character(K1),as.character(K2)) )
typeof(visitor_reward[,1])
## Change type
visitor_reward <- ChangeType(visitor_reward)
typeof(visitor_reward[,1])
}
