% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_epsilon_greedy.R
\name{condition_for_epsilon_greedy}
\alias{condition_for_epsilon_greedy}
\title{Condition for epsilon greedy}
\usage{
condition_for_epsilon_greedy(S, epsilon = 0.25)
}
\arguments{
\item{S}{Numeric matrix of means and trials}

\item{epsilon}{Numeric value (optional)}
}
\value{
Integer value
}
\description{
Chooses the best with 1-epsilon probability. A random arm with
 probability epsilon.
}
\examples{
## Generates 1000 numbers from 2 binomial distributions
set.seed(4434)
K1 <- rbinom(1000, 1, 0.6)
K2 <- rbinom(1000, 1, 0.7)
## Define a dataframe of rewards
visitor_reward <- as.data.frame( cbind(K1,K2) )
## Number of arms
K=2
## Init the S Matrix
S <- generate_matrix_S(K)
S
## play arms uniformly
for(i in 1:nrow(visitor_reward)){
S <- play_arm(i,arm=(i\%\%K+1),S,visitor_reward)
}
## Results
S

}
