% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_exp3.R
\name{condition_for_exp3}
\alias{condition_for_exp3}
\title{Condition for EXP3}
\usage{
condition_for_exp3(K, proba)
}
\arguments{
\item{K}{Number of arms}

\item{proba}{Vector of probability weights}
}
\value{
Integer value
}
\description{
Samples the arm list according to a vector of probability weights
}
