% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_kl_ucb.R
\name{condition_for_klucb}
\alias{condition_for_klucb}
\title{Arm indices according to the Kullback-Leibler upper confidence bound}
\usage{
condition_for_klucb(S, d, precision, max_iteration)
}
\arguments{
\item{S}{Means/Trials matrix}

\item{d}{Upper bound of the divergence}

\item{precision}{Precision of bisection method}

\item{max_iteration}{Maximum number of iterations}
}
\value{
A vector containing the arms indices
}
\description{
Computes for each arm a the value m (assumed to be the mean of a
  binomial distribution) that maximize {kl(mu_hat(a),m) < d} where kl is the
  binomial Kullback-Leibler divergence and mu_hat(a) the reward expectation
  of arm a. The value is computed using a bisection method.
}
