% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_for_epsilon_greedy.R
\name{ConditionForEpsilonGreedy}
\alias{ConditionForEpsilonGreedy}
\title{ConditionForEpsilonGreedy}
\usage{
ConditionForEpsilonGreedy(S, epsilon = 0.25, K = ncol(S))
}
\arguments{
\item{S}{Numeric matrix}

\item{epsilon}{Numeric value (optional)}

\item{K}{Integer value (optional)}
}
\value{
Integer value
}
\description{
choose the best with 1 - espsilon probability. 1 : best arm , 2 : other arm
}
\examples{
## Generates 1000 numbers from 2 binomial distributions
set.seed(4434)
K1 <- rbinom(1000, 1, 0.6)
K2 <- rbinom(1000, 1, 0.7)
## Define a dataframe of rewards
visitor_reward <- as.data.frame( cbind(K1,K2) )
## Number of arms
K=2
## Init the S Matrix
S <- GenerateMatrixS(K)
S
## play arms uniformly
for(i in 1:nrow(visitor_reward)){
S <- PlayArm(i,arm=(i\%\%K+1),S,visitor_reward)
}
## Results
S
ConditionForEpsilonGreedy(S=S)
}
