% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit_reward_control.R
\name{control_binary}
\alias{control_binary}
\title{Are rewards binary values ?}
\usage{
control_binary(visitor_reward)
}
\arguments{
\item{visitor_reward}{Numeric matrix}
}
\value{
Logical value
}
\description{
Checks if reward matrix contains only numeric binary values.
 Stops if not. Returns TRUE otherwise.
}
