% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_data_missing.R
\name{control_data_missing}
\alias{control_data_missing}
\title{control_data_missing}
\usage{
control_data_missing(visitorReward)
}
\arguments{
\item{visitorReward}{Dataframe of integer or numeric values}
}
\value{
Logical value
}
\description{
Control data for bandit.
Check in a dataframe if there is some missing values
Print a message if it's not respected.
Else return TRUE.
}
\examples{
## Generates 1000 numbers from 2 uniform distributions
set.seed(4434)
K1 <- rnorm(100, 30, .05)
K2 <- rnorm(100, 21, .05)
## Define a dataframe of rewards
visitorReward <- as.data.frame(cbind(K1,K2) )
## Control
control_data_missing(visitorReward)
visitorReward[1,1]= NA
## Control
control_data_missing(visitorReward)

}
