% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit_reward_control.R
\name{control_missing_data}
\alias{control_missing_data}
\title{Missing data control of visitor reward dataframe}
\usage{
control_missing_data(visitor_reward)
}
\arguments{
\item{visitor_reward}{Dataframe of integer or numeric values}
}
\value{
Logical value
}
\description{
Controls data for bandit. Check in a dataframe if there is some
 missing values. Stops if it's not respected. Else returns TRUE.
}
