% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit_reward_control.R
\name{control_size}
\alias{control_size}
\title{Size control of visitor reward dataframe}
\usage{
control_size(visitor_reward)
}
\arguments{
\item{visitor_reward}{Dataframe of numeric values}
}
\value{
Logical value
}
\description{
Controls number of arms for bandit. Check if the number of
 columns is greater than or equal to 2 and if the number of rows is greater
 than or equal to the number of columns. Stops if not respected. Else returns
 TRUE.
}
