% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_update_modulo.R
\name{control_update_modulo}
\alias{control_update_modulo}
\title{control_update_modulo}
\usage{
control_update_modulo(choice, M, update_val)
}
\arguments{
\item{choice}{an integer value associated with the selected arm}

\item{M}{Dataframe of integer or numeric values}

\item{update_value}{first item used for cumpute the regression;}
}
\value{
a new model
}
\description{
Check if the model has to be re-compute (number of item equals to update_val have been tested)
Return TRUE if we need to re cumpute the model
}
\examples{
size.tot = 1000
set.seed(4649)                          # this makes the example exactly reproducible
x1 = runif(size.tot, min=0, max=10)          # you have 4, largely uncorrelated predictors
x2 = runif(size.tot, min=0, max=10)
x3 = runif(size.tot, min=0, max=10)
x4 = runif(size.tot, min=0, max=10)
dt = cbind(x1,x2,x3,x4)
arm_1 <-  as.vector(c(-1,9,-8,4))
K1 = crossprod(t(dt),arm_1)
dt <-as.data.frame(dt)
M <- dt
M$reward <- K1
M$arm <- 0
M$arm[1:99] <- 1
control_update_modulo(1, M, 100)
M$arm[100] <- 1
control_update_modulo(1, M, 100)
}
