% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createClusters.R
\name{createClusters}
\alias{createClusters}
\title{createClusters}
\usage{
createClusters(
  listSerie = colnames(dt),
  dt,
  method = "DBA",
  listKCentroids,
  plotCentroids = TRUE,
  plotClusters = TRUE
)
}
\arguments{
\item{listSerie}{List of characters (Colom Name)}

\item{dt}{Dataframe of time series}

\item{method}{Method for clustering (DBA or HCLUST), DBA is set by default}

\item{listKCentroids}{List of Integer (Number of centroid)}

\item{plotCentroids}{Boolean (optional) If you want to see the centroids}

\item{plotClusters}{Boolean (optional) If you want to see the clusters}
}
\value{
Integer value
}
\description{
Create cluster based on a specified clustering method
}
\examples{
set.seed(100000)
alpha_list <- c(1,2,3)
beta_list <- c(0.3,0.2,-0.2)
theta_list <- c(0.8,0.2,0.5)
y <- as.data.frame(c(1))
colnames(y) = "ID"
temp=1
for (j in 1:10){
 for (i in 1:length(alpha_list)){
   n = sample(1:1000,1)
   t <- 1:n
   ts <- alpha_list[i] + beta_list[i] * t + arima.sim(list(ma = theta_list[i]), n = length(t))
   y[temp, "time_series"][[1]] <- list(ts)
   y[temp, "cluster"][[1]] <- i
   y$ID[temp] = temp
   temp = temp +1
 }
}
obj <- createClusters(listSerie = c("time_series") , dt = y , method = "DBA" , listKCentroids=c(3) , plotCentroids = TRUE , plotClusters = TRUE )
}
