% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctree_formula_generate.R
\name{ctree_formula_generate}
\alias{ctree_formula_generate}
\title{ctree_formula_generate}
\usage{
ctree_formula_generate(
  dt,
  visitor_reward,
  arm_for_learn,
  explanatory_variable,
  learn_size,
  ctree_control_val,
  print = FALSE
)
}
\arguments{
\item{dt}{Dataframe of integer numeric or factor values}

\item{visitor_reward}{Dataframe of integer or numeric values}

\item{arm_for_learn}{arm dedicated to the learnset (step 1) (optional),}

\item{explanatory_variable}{= list of covariates (optional),}

\item{learn_size}{number of items dedicated to the learnset (step 1) (optional),}

\item{ctree_control_val}{= Various parameters that control aspects of the ‘ctree’ fit (optional),}

\item{is_reward_are_boolean}{logical value (optional)}
}
\value{
An object of class \code{\link{party}}
}
\description{
Generate a conditional inference tree with \code{\link{ctree}} function.
Apply a aecursive partitioning for continuous, censored, ordered, nominal and multivariate response variables in a conditional inference framework.
}
\examples{
size.tot = 1000
set.seed(4649)                          # this makes the example exactly reproducible
x1 = runif(size.tot, min=0, max=10)          # you have 4, largely uncorrelated predictors
x2 = runif(size.tot, min=0, max=10)
x3 = runif(size.tot, min=0, max=10)
x4 = runif(size.tot, min=0, max=10)
dt = cbind(x1,x2,x3,x4)
arm_1 <-  as.vector(c(-1,9,-8,4))
K1 = crossprod(t(dt),arm_1)
arm_2 <-  as.vector(c(-1,2,1,0))
K2 = crossprod(t(dt),arm_2)
arm_3 <-  as.vector(c(-1,-5,1,10))
K3 = crossprod(t(dt),arm_3)
visitor_reward <-  data.frame(K1,K2,K3)
dt <- as.data.frame(dt)
ctreeucb_control_val=ctreeucb_parameters_control_default(dt=dt,  visitor_reward= visitor_reward)
ctree_formula_generate(dt, visitor_reward,arm_for_learn='K1',explanatory_variable=c('x1','x2','x3','x4'),learn_size=100,ctree_control_val=ctreeucb_control_val$ctree_control_val)
ctree_formula_generate(dt, visitor_reward,arm_for_learn='K2',explanatory_variable=c('x1','x2','x3','x4'),learn_size=100,ctree_control_val=ctreeucb_control_val$ctree_control_val)
ctree_formula_generate(dt, visitor_reward,arm_for_learn='K3',explanatory_variable=c('x1','x2','x3','x4'),learn_size=100,ctree_control_val=ctreeucb_control_val$ctree_control_val)
}
