% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctreeucb_parameters_control_default.R
\name{ctreeucb_parameters_control_default}
\alias{ctreeucb_parameters_control_default}
\title{ctreeucb_parameters_control_default}
\usage{
ctreeucb_parameters_control_default(
  dt,
  visitor_reward,
  is_reward_are_boolean = FALSE,
  learn_size = as.integer(nrow(dt) * 0.1),
  arm_for_learn = names(visitor_reward)[1],
  explanatory_variable = names(dt),
  alpha = 1,
  ctree_control_val = ctree_control()
)
}
\arguments{
\item{dt}{Dataframe of integer numeric or factor values}

\item{visitor_reward}{Dataframe of integer or numeric values}

\item{is_reward_are_boolean}{logical value (optional) (optional),}

\item{learn_size}{number of items dedicated to the learnset (step 1) (optional),}

\item{arm_for_learn}{arm dedicated to the learnset (step 1) (optional),}

\item{explanatory_variable}{=  list of covariates (optional),}

\item{alpha}{Numeric value (optional)}

\item{ctree_control_val}{Various parameters that control aspects of the ‘ctree’ fit (optional),}
}
\value{
list of parameters
}
\description{
Various parameters that control aspects of the ctreeUCB fit. See also \code{\link{ctree_control}} and \code{\link{ctree}}
}
\examples{
c1 <- rep("Test",100)
dt <- as.data.frame(c1)
K1 <- sample(c(0,1),replace=TRUE,size= 100)
visitor_reward <- as.data.frame(K1)
ctreeucb_parameters_control_default(dt=dt, visitor_reward=visitor_reward)
ctreeucb_parameters_control_default(dt=dt, visitor_reward=visitor_reward,is_reward_are_boolean=TRUE,learn_size=200, explanatory_variable="c1", alpha=1, ctree_control_val= partykit::ctree_control(teststat = "quadratic"))
}
