% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_regret.R
\name{cumulativeRegret}
\alias{cumulativeRegret}
\title{Return list of cumulative regret}
\usage{
cumulativeRegret(choice, visitor_reward)
}
\arguments{
\item{choice}{Integer list}

\item{visitor_reward}{dataframe of integer or numeric values}
}
\value{
List of numeric values
}
\description{
Plot the cumulative regret overt the time.
Return a list with cumulative regret at each iterations
}
\examples{
## Generates 1000 numbers from 2 uniform distributions
set.seed(4434)
K1 <- rnorm(100, 30, .05)
K2 <- rnorm(100, 21, .05)
## Define a dataframe of rewards
visitor_reward <- as.data.frame(cbind(K1,K2) )
#Random choices
choice <- sample(c(1,2), 100, replace = TRUE)
CumulativeRegret(choice=choice,visitor_reward=visitor_reward)

}
