% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_control_context_reward.R
\name{data_control_context_reward}
\alias{data_control_context_reward}
\title{data_control_context_reward}
\usage{
data_control_context_reward(dt, visitorReward,
  is_reward_are_boolean = FALSE)
}
\arguments{
\item{dt}{Dataframe of integer numeric or factor values}

\item{visitorReward}{Dataframe of integer or numeric values}

\item{is_reward_are_boolean}{logical value (optional)}
}
\value{
Logical value
}
\description{
Control if number of item in data reward and context data are equal
Check if a reward is defined as logical value.
Print a message and stop if this two conditions are not respected.
Else return TRUE.
}
\examples{
## Generates 1000 numbers from 2 uniform distributions
set.seed(4434)
K1 <- rnorm(100, 30, .05)
K2 <- rnorm(100, 35, .05)
## Define a dataframe of rewards
visitorReward <- as.data.frame(K1)
## Define a dataframe of context
c1 <- rnorm(50, 35, .05)
dt <- as.data.frame(c1)
## Control
data_control_context_reward(dt=dt,visitorReward=visitorReward)
c1 <- rep("Test",100)
dt <- as.data.frame(c1)
data_control_context_reward(dt=dt,visitorReward=visitorReward,is_reward_are_boolean =TRUE)
c1 <- rnorm(100, 30, .05)
dt <- as.data.frame(c1)
data_control_context_reward(dt=dt,visitorReward=visitorReward,is_reward_are_boolean =TRUE)
data_control_context_reward(dt=dt,visitorReward=visitorReward)

}
