% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_control_K.R
\name{data_control_K}
\alias{data_control_K}
\title{data_control_K}
\usage{
data_control_K(visitorReward, K = ncol(visitorReward))
}
\arguments{
\item{visitorReward}{Dataframe of integer or numeric values}

\item{K}{Integer value (optional)}
}
\value{
Logical value
}
\description{
Control arm and data for bandit
Check if a dataframe gets an equal number of colonms than K possible arms.
Check if K \geq 2. Print a message  and stop if this two conditions are not respected.
Else return TRUE.
}
\examples{
## Generates 1000 numbers from 2 uniform distributions
set.seed(4434)
K1 <- rnorm(100, 30, .05)
## Define a dataframe of rewards
visitorReward <- as.data.frame(K1)
## Control
data_control_K(visitorReward)
K2 <- rnorm(100, 21, .05)
## Define a dataframe of rewards
visitorReward <- as.data.frame(cbind(K1,K2) )
## Control
data_control_K(visitorReward,K=3)
## Control
data_control_K(visitorReward,K=2)

}
