% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_control_context_reward.R
\name{DataControlContextReward}
\alias{DataControlContextReward}
\title{DataControlContextReward}
\usage{
DataControlContextReward(dt, visitor_reward)
}
\arguments{
\item{dt}{Dataframe of integer numeric or factor values}

\item{visitor_reward}{Dataframe of integer or numeric values}
}
\value{
Logical value
}
\description{
Control if number of item in data reward and context data are equal
Print a message and stop if this condition is not respected.
Else return TRUE.
}
\examples{
## Generates 1000 numbers from 2 uniform distributions
set.seed(4434)
K1 <- rnorm(100, 30, .05)
K2 <- rnorm(100, 35, .05)
## Define a dataframe of rewards
visitor_reward <- as.data.frame(K1)
## Define a dataframe of context
c1 <- rnorm(100, 30, .05)
dt <- as.data.frame(c1)
DataControlContextReward(dt=dt,visitor_reward=visitor_reward)

}
