% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_control_K.R
\name{DataControlK}
\alias{DataControlK}
\title{DataControlK}
\usage{
DataControlK(visitor_reward, K = ncol(visitor_reward))
}
\arguments{
\item{visitor_reward}{Dataframe of integer or numeric values}

\item{K}{Integer value (optional)}
}
\value{
Logical value
}
\description{
Control arm and data for bandit
Check if a dataframe gets an equal number of colonms than K possible arms.
Check if K geq 2. Print a message  and stop if this two conditions are not respected.
Else return TRUE.
}
\examples{
## Generates 1000 numbers from 2 uniform distributions
set.seed(4434)
K1 <- rnorm(100, 30, .05)
## Define a dataframe of rewards
visitor_reward <- as.data.frame(K1)
## Control
#DataControlK(visitor_reward)
K2 <- rnorm(100, 21, .05)
## Define a dataframe of rewards
visitor_reward <- as.data.frame(cbind(K1,K2) )
## Control
DataControlK(visitor_reward,K=2)

}
