% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/dbalinucbRejectionSamplingBanditObjectEvaluation.R
\name{dbalinucbRejectionSamplingBanditObjectEvaluation}
\alias{dbalinucbRejectionSamplingBanditObjectEvaluation}
\title{dbalinucbRejectionSamplingBanditObjectEvaluation}
\usage{
dbalinucbRejectionSamplingBanditObjectEvaluation(
  dt,
  visitor_reward,
  alpha = 1,
  K = ncol(visitor_reward),
  listSerie,
  listKCentroids,
  learn_size = as.integer(nrow(dt) * 0.3),
  IsRewardAreBoolean = FALSE,
  listCategorial = 0,
  listInteger = 0
)
}
\arguments{
\item{dt}{Dataframe of integer or numeric values}

\item{visitor_reward}{Dataframe of integer or numeric values}

\item{alpha}{numerical value (optional)}

\item{K}{Integer value (optional)}

\item{average}{Boolean value to evaluate the policy}
}
\value{
\itemize{ List of element:
 \item dbalinucb_rejection_sampling_alloc: dbalinucb bandit object ,
 \item cum_reg_dbalinucb_bandit_alloc: List numeric.
 }
}
\description{
Run a \code{\link{DBALINUCB_rejection_sampling}} using visitor_reward and dt values.
Control data.
Stop if something is wrong.
Exclud any choices which not corresponds to real exepriments in dataset.
After execution of DBALINUCB_rejection_sampling, calculates the cumulative regret
associated with the choices made.
Review the cumulative regret according iterations and an DBALINUCB_rejection_sampling object.
See also  \code{\link{DBALINUCB_rejection_sampling}}
Require \code{\link{tic}} and \code{\link{toc}} from \code{\link{tictoc}} library
}
\examples{
size.tot = 9000
}
