% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsilon_greedy_bandit_object_evaluation.R
\name{EpsilonGreedyBanditObjectEvaluation}
\alias{EpsilonGreedyBanditObjectEvaluation}
\title{EpsilonGreedyBanditObjectEvaluation}
\usage{
EpsilonGreedyBanditObjectEvaluation(
  visitor_reward = visitor_reward,
  K = ncol(visitor_reward),
  epsilon = 0.25,
  average = FALSE,
  IsRewardAreBoolean = TRUE,
  dt = NA,
  explanatory_variable = colnames(dt)
)
}
\arguments{
\item{visitor_reward}{Dataframe of integer or numeric values}

\item{K}{Integer value (optional)}

\item{epsilon}{Numeric value (optional)}

\item{average}{Boolean value (optional)}
}
\value{
\itemize{ List of element:
 \item EpsilonGreedy_alloc: EpsilonGreedy object ,
 \item cum_reg_EpsilonGreedy_alloc: List numeric.
 }
}
\description{
Run the EpsilonGreedy algorithm using visitor_reward values with \code{\link{EpsilonGreedy}} function.
Stop if something is wrong.
After execution of EpsilonGreedy, calculates the cumulative regret
associated with the choices made.
Review the cumulative regret according iterations and an EpsilonGreedy object.
See also \code{\link{EpsilonGreedy}}, \code{\link{CumulativeRegret}}
Require \code{\link{tic}} and \code{\link{toc}} from \code{\link{tictoc}} library
}
\examples{
## Generates 1000 numbers from 2 binomial distributions
set.seed(4434)
K1 <- rbinom(1000, 1, 0.6)
K2 <- rbinom(1000, 1, 0.7)
## Define a dataframe of rewards
visitor_reward <- as.data.frame(cbind(K1,K2) )
#Run EpsilonGreedy algorithm with policy evaluation
EpsilonGreedyBanditObjectEvaluation(visitor_reward,epsilon = 0.25)

}
