% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_matrix_S.R
\name{generate_matrix_S}
\alias{generate_matrix_S}
\title{Generate a S matrix}
\usage{
generate_matrix_S(K)
}
\arguments{
\item{K}{Integer variable. Number of arms}
}
\value{
Numeric matrix
}
\description{
Generates the matrix used to store UCB computation variables.
  This is a 2 by K (number of arms) matrix with each columns corresponding to
  an arm. First row will store the estimated reward expectation, second row
  the number of times the arm was played.
}
\examples{
K = 2
generate_matrix_S(K)

}
