% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_matrix_S.R
\name{GenerateMatrixS}
\alias{GenerateMatrixS}
\title{Generate a S Matrix}
\usage{
GenerateMatrixS(x)
}
\arguments{
\item{x}{Integer variable}
}
\value{
Numeric matrix
}
\description{
Returns a matrix initialized to 0 of dimension K by 2.
Allows to save the average empirical reward (first line) and number of trials (second line) of each arm into a matrix S.
Require a numercial values to define the number of possible arm
}
\examples{
K = 2
GenerateMatrixS(K)

}
