% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_reward_are_boolean.R
\name{IsRewardAreBoolean}
\alias{IsRewardAreBoolean}
\title{IsRewardAreBoolean}
\usage{
IsRewardAreBoolean(visitor_reward)
}
\arguments{
\item{visitor_reward}{Dataframe of integer or numeric values}
}
\value{
Logical value
}
\description{
Check if a reward is defined as logical value.
Print a message and stop if this condition is not respected.
Else return TRUE.
}
\examples{
## Generates 1000 numbers from 2 uniform distributions
set.seed(4434)
K1 <- rbinom(1000, 1, 0.6)
K2 <- rbinom(1000, 1, 0.7)
visitor_reward <- as.data.frame(cbind(K1,K2))
IsRewardAreBoolean(visitor_reward)

}
