% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_kl_ucb.R
\name{kl_bernoulli}
\alias{kl_bernoulli}
\title{Kullback-Leibler divergence of two Bernoulli distributions}
\usage{
kl_bernoulli(p, q)
}
\arguments{
\item{p}{Mean of first Bernouilli distribution}

\item{q}{Mean of second Bernouilli distribution}
}
\value{
The Kullback-Leibler divergence of the two distributions
}
\description{
The Kullback-Leibler divergence is a measure of how different
  two distributions are. Here two Bernouilli distributions parameterized by
  their means p and q.
}
