% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL_UCB.R
\name{KLUCB}
\alias{KLUCB}
\title{KL-UCB algorithm}
\usage{
KLUCB(visitor_reward, K = ncol(visitor_reward), precision = 1e-06, c = 0)
}
\arguments{
\item{visitor_reward}{Dataframe of integer or numeric values}

\item{K}{Integer value (optional)}

\item{precision}{Numeric value (optional)}

\item{c}{Numeric value (optional)}
}
\value{
\itemize{ List of element:
 \item S:numerical matrix of results ,
 \item choice: choices of KL-UCB,
 \item proba: probability of the chosen arms,
 \item time: time of cumputation,
 \item theta_hat: mean estimated of each arm
 \item theta: real mean of each arm
 }
}
\description{
Kullback-Leibler Upper Confidence Bound (KL-UCB) bandit strategy.
Generate a matrix to save the results (S).
Returns the calculation time.
Review the estimated, actual averages and number of choices for each arm.
See also  \code{\link{ConditionForKLUCB}}, \code{\link{kl_bernoulli}}, \code{\link{kl_ucb_bernoulli}},
\code{\link{kl_ucb_gaussian}}, \code{\link{GenerateMatrixS}}, and \code{\link{PlayArm}}.
Require \code{\link{tic}} and \code{\link{toc}} from \code{\link{tictoc}} library
}
\examples{
## Generates 1000 numbers from 2 uniform distributions
set.seed(4434)
K1 <- rbinom(1000, 1, 0.6)
K2 <- rbinom(1000, 1, 0.7)
## Define a dataframe of rewards
visitor_reward <- as.data.frame( cbind(K1,K2) )
klucb_alloc <- KLUCB(visitor_reward)
klucb_alloc$S
klucb_alloc$time
klucb_alloc$theta
klucb_alloc$theta_hat
}
