% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listKCentroidsControl.R
\name{listKCentroidsControl}
\alias{listKCentroidsControl}
\title{listKCentroidsControl}
\usage{
listKCentroidsControl(listKCentroids, listSeriesList)
}
\arguments{
\item{listSeriesList}{List of integer vector}
}
\value{
Logical value
}
\description{
Control if the listSeries and listKCentroidsControl corresponds
Print a message and stop if this condition is not respected.
Else return TRUE.
}
\examples{
listKCentroids = c(1,2,3)
listSeriesList = c("Type 1","Type 2")
listKCentroidsControl(listKCentroids, listSeriesList)
}
