% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOGITUCB.R
\name{LOGITUCB}
\alias{LOGITUCB}
\title{LogitUCB algorithm}
\usage{
LOGITUCB(dt, visitorReward, alpha = 1, K = ncol(visitorReward))
}
\arguments{
\item{dt}{Dataframe of integer or numeric values}

\item{visitorReward}{Dataframe of integer or numeric values}

\item{alpha}{Numeric value (optional)}

\item{K}{Integer value (optional)}
}
\value{
\itemize{ List of element:
 \item choice: choices of UCB,
 \item proba: probability of the chosen arms,
 \item time: time of cumputation,
 \item theta_hat: coefficients estimated of each arm
 \item theta: real coefficients of each arm
 }
}
\description{
Control data in visitorReward with \code{\link{bandit_reward_control}}
Stop if something is wrong.
\itemize{ At each iteration
 \item Calculates the arm probabilities according to a logit regression of context in dt dataframe
 \item Choose the arm with the maximum upper bound (with \alpha parameter)
 \item Receives a reward in visitorReward for the arm and associated iteration
 \item Updates the results matrix S.
 }
Returns the calculation time.
Review the estimated, actual coefficient for each arm.
See also  \code{\link{return_real_theta}},
Require \code{\link{tic}} and \code{\link{toc}} from \code{\link{tictoc}} library
}
\examples{
size.tot = 1000
set.seed(4649)                          # this makes the example exactly reproducible
x1 = runif(size.tot, min=0, max=10)          # you have 4, largely uncorrelated predictors
x2 = runif(size.tot, min=0, max=10)
x3 = runif(size.tot, min=0, max=10)
x4 = runif(size.tot, min=0, max=10)
dt = cbind(x1,x2,x3,x4)
#arm reward
arm_1 <-  as.vector(c(-1,9,-8,4))
K1 = 1/(1+exp(- crossprod(t(dt),arm_1))) # inverse logit transform of linear predictor
K1 = vapply(K1, function(x) rbinom(1, 1, x), as.integer(1L))
arm_2 <-  as.vector(c(-1,2,1,0))
K2 = 1/(1+exp(- crossprod(t(dt),arm_2))) # inverse logit transform of linear predictor
K2 = vapply(K2, function(x) rbinom(1, 1, x), as.integer(1L))
arm_3 <-  as.vector(c(-1,-5,1,10))
K3 = 1/(1+exp(- crossprod(t(dt),arm_3)))
K3 = vapply(K3, function(x) rbinom(1, 1, x), as.integer(1L))
visitorReward <-  data.frame(K1,K2,K3)
dt <- as.data.frame(dt)
LOGITUCB(dt,visitorReward)
}
