% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_ucb.R
\name{proba_max_for_ucb}
\alias{proba_max_for_ucb}
\title{Upper Confidence Bound computation function}
\usage{
proba_max_for_ucb(S, iter, alpha = 1)
}
\arguments{
\item{S}{Means and trials matrix}

\item{iter}{Numeric value. Index of iteration.}

\item{alpha}{Numeric value. Exploration parameter (optional)}
}
\value{
Numeric vector containing each arm upper confidence bound.
}
\description{
Computes, for each arm an upper bound according to the Hoeffding
 inequality parameterized by exploration parameter alpha (default alpha = 1).
 Returns a vector containing each arm upper confidence bound.
}
\examples{
## Generates 1000 numbers from 2 uniform distributions
set.seed(4434)
K1 <- rbinom(1000, 1, 0.6)
K2 <- rbinom(1000, 1, 0.7)
## Define a dataframe of rewards
visitor_reward <- as.data.frame( cbind(K1,K2) )
## Init the S Matrix
S <- generate_matrix_S(K=2)
S
## play arms uniformly
for(i in 1:nrow(visitor_reward)){
S <- play_arm(i,arm=(i\%\%K+1),S,visitor_reward)
}
## Results
S
proba_max_for_ucb(S=S,iter=i+1)

}
