% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proba_max_for_UCB.R
\name{ProbaMaxForUCB}
\alias{ProbaMaxForUCB}
\title{Calculates, for each colonne of S (selected arm) an
upper bound according to the Hoeffding inequality
(dependent on the iter iteration).
It is possible to adjust this bound via an
alpha parameter (default alpha = 1).
Returns a vector of calculated upper bounds}
\usage{
ProbaMaxForUCB(S, iter, alpha = 1, K = ncol(S))
}
\arguments{
\item{S}{Numeric matrix}

\item{iter}{Integer value}

\item{alpha}{Numeric value (optional)}

\item{K}{Integer value (optional)}
}
\value{
Numeric vector
}
\description{
Calculates, for each colonne of S (selected arm) an
upper bound according to the Hoeffding inequality
(dependent on the iter iteration).
It is possible to adjust this bound via an
alpha parameter (default alpha = 1).
Returns a vector of calculated upper bounds
}
\examples{
## Generates 1000 numbers from 2 uniform distributions
set.seed(4434)
K1 <- rbinom(1000, 1, 0.6)
K2 <- rbinom(1000, 1, 0.7)
## Define a dataframe of rewards
visitor_reward <- as.data.frame( cbind(K1,K2) )
## Number of arms
K=2
## Init the S Matrix
S <- GenerateMatrixS(K)
S
## play arms uniformly
for(i in 1:nrow(visitor_reward)){
S <- PlayArm(i,arm=(i\%\%K+1),S,visitor_reward)
}
## Results
S
ProbaMaxForUCB(S=S,iter=i+1)
}
