% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regretValue.R
\name{regretValue}
\alias{regretValue}
\title{Return regret of chosen arm}
\usage{
regretValue(arm, vec_visitorReward)
}
\arguments{
\item{arm}{Integer value}

\item{vec_visitorReward}{Numeric vector}
}
\value{
Numeric value
}
\description{
Return for a given vector the difference between
the highest value of reward and the reward obtained by the chosen arm.
Can be equal of 0
}
\examples{
## Generates 1000 numbers from 2 uniform distributions
set.seed(4434)
K1 <- rnorm(100, 30, .05)
K2 <- rnorm(100, 21, .05)
## Define a dataframe of rewards
visitorReward <- as.data.frame(cbind(K1,K2) )
#regret of arm 1 for the fist item
regretValue(1,visitorReward[1,])
#'#regret of arm 1 for the fist item
regretValue(2,visitorReward[1,])
}
