% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_variance_error_Kernelucb.R
\name{return_variance_error_Kernelucb}
\alias{return_variance_error_Kernelucb}
\title{return_variance_error_Kernelucb}
\usage{
return_variance_error_Kernelucb(choice, expected, rewards)
}
\arguments{
\item{choice}{an integer value associated with the selected arm}

\item{expected}{vector of numeric values expected according to a selected arm}

\item{rewards}{vector of numeric values obtained according to a selected arm}
}
\value{
variance of error
}
\description{
Cumpute variance of a kernelized regression models with \code{\link{krr}} function (library listdtr)
based on selected data.
}
\examples{
size.tot = 1000
set.seed(4649)                          # this makes the example exactly reproducible
x1 = runif(size.tot, min=0, max=10)          # you have 4, largely uncorrelated predictors
x2 = runif(size.tot, min=0, max=10)
x3 = runif(size.tot, min=0, max=10)
x4 = runif(size.tot, min=0, max=10)
dt = cbind(x1,x2,x3,x4)
#arm reward
arm_1 <-  as.vector(c(-1,9,-8,4))
K1 = crossprod(t(dt),arm_1)
#covariates
dt <-as.data.frame(dt)
#prepare data for regression
M <- dt
M$arm <- 1
M$reward <- K1
#cumpute the model
ker_model <- update_model_for_KernelUCB(choice=1 , M=M ,update_val=100)
#evaluate the model
xnew <- as.data.frame(dt)
predictions <- predict(ker_model$model, xnew)
return_variance_error_Kernelucb(choice=M$arm ,expected=predictions, rewards=K1)
}
