% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_real_theta.R
\name{ReturnRealTheta}
\alias{ReturnRealTheta}
\title{ReturnRealTheta}
\usage{
ReturnRealTheta(dt, visitor_reward, option = "linear")
}
\arguments{
\item{dt}{Dataframe of context}

\item{visitor_reward}{Dataframe of integer or numeric values}

\item{option}{Option linear by default}
}
\value{
theta_hat: mean estimated of each arm
}
\description{
Return real theta from a rigide regression of context to arm's reward.
Return coefficients of regression (except intercept)
See also , \code{\link{lm}}.
}
\examples{
size.tot = 1000
set.seed(4649)                          # this makes the example exactly reproducible
x1 = runif(size.tot, min=0, max=10)          # you have 4, largely uncorrelated predictors
x2 = runif(size.tot, min=0, max=10)
x3 = runif(size.tot, min=0, max=10)
x4 = runif(size.tot, min=0, max=10)
dt = cbind(x1,x2,x3,x4)
#arm reward
arm_1 <-  as.vector(c(-1,9,-8,4))
K1 = crossprod(t(dt),arm_1)
arm_2 <-  as.vector(c(-1,2,1,0))
K2 = crossprod(t(dt),arm_2)
arm_3 <-  as.vector(c(-1,-5,1,10))
K3 = crossprod(t(dt),arm_3)
visitor_reward <-  data.frame(K1,K2,K3)
dt <- as.data.frame(dt)
ReturnRealTheta(dt,visitor_reward)
}
