% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_categorial_to_binary.R
\name{transform_categorial_to_binary}
\alias{transform_categorial_to_binary}
\title{transform_categorial_to_binary}
\usage{
transform_categorial_to_binary(listCategorial = 0, listInteger = 0, dt)
}
\arguments{
\item{listCategorial}{list of categorial columns}

\item{listInteger}{list of numerical columns}

\item{dt}{Dataframe of integer or numeric values}
}
\value{
A dataframe with numerical value
}
\description{
transform categorial columns into binary
}
