% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_bandit.R
\name{UniformBandit}
\alias{UniformBandit}
\title{Uniform algorithm}
\usage{
UniformBandit(visitor_reward, K = ncol(visitor_reward))
}
\arguments{
\item{visitor_reward}{Dataframe of integer or numeric values}

\item{K}{Integer value (optional)}
}
\value{
\itemize{ List of element:
 \item S:numerical matrix of results ,
 \item choice: choices of UCB,
 \item time: time of cumputation,
 }
}
\description{
Control data in visitor_reward with \code{\link{BanditRewardControl}}
Stop if something is wrong.
Generate a matrix to save the results (S).
\itemize{ At each iteration
 \item Choose alternatively an arm
 \item Receives a reward in visitor_reward for the arm and associated iteration
 \item Updates the results matrix S.
 }
Returns the calculation time.
Return the estimated and number of choices for each arm.
See also  \code{\link{GenerateMatrixS}} and \code{\link{PlayArm}}.
Require \code{\link{tic}} and \code{\link{toc}} from \code{\link{tictoc}} library
}
\examples{
## Generates 10000 numbers from 2 binomial  distributions
set.seed(4434)
K1 <- rbinom(1000, 1, 0.6)
K2 <- rbinom(1000, 1, 0.7)
## Define a dataframe of rewards
visitor_reward <- as.data.frame(cbind(K1,K2) )
#Run Uniform algorithm
uniform_bandit_alloc  <- UniformBandit(visitor_reward)
uniform_bandit_alloc$S
uniform_bandit_alloc$time
}
