% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniformBanditRejectionSampling.R
\name{UniformBanditRejectionSampling}
\alias{UniformBanditRejectionSampling}
\title{Uniform Rejection Sampling algorithm}
\usage{
UniformBanditRejectionSampling(visitorReward, K = ncol(visitor_reward))
}
\arguments{
\item{K}{Integer value (optional)}

\item{visitor_reward}{Dataframe of integer or numeric values}
}
\value{
\itemize{ List of element:
 \item S:numerical matrix of results ,
 \item choice: choices of UCB,
 \item time: time of cumputation,
 }
}
\description{
Control data in visitor_reward with \code{\link{BanditRewardControl}}
Stop if something is wrong.
Generate a matrix to save the results (S).
\itemize{ At each iteration
 \item Choose alternatively an arm
 \item Receives a reward in visitor_reward for the arm and associated iteration
 \item Updates the results matrix S.
 }
Returns the calculation time.
Return the estimated and number of choices for each arm.
See also  \code{\link{GenerateMatrixS}} and \code{\link{PlayArm}}.
Require \code{\link{tic}} and \code{\link{toc}} from \code{\link{tictoc}} library
}
\examples{
## TODO
}
