% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_model_for_KernelUCB.R
\name{update_model_for_KernelUCB}
\alias{update_model_for_KernelUCB}
\title{update_model_for_KernelUCB}
\usage{
update_model_for_KernelUCB(choice, M, update_val)
}
\arguments{
\item{choice}{an integer value associated with the selected arm}

\item{M}{Dataframe of integer or numeric values}

\item{update_value}{first item used for cumpute the regression;}
}
\value{
a new model
}
\description{
Cumpute a kernelized regression models with \code{\link{krr}} function (library listdtr)
on selected data.
This function require a lot of time so an uptate value can be use for limit the number of used data.
}
\examples{
size.tot = 1000
set.seed(4649)                          # this makes the example exactly reproducible
x1 = runif(size.tot, min=0, max=10)          # you have 4, largely uncorrelated predictors
x2 = runif(size.tot, min=0, max=10)
x3 = runif(size.tot, min=0, max=10)
x4 = runif(size.tot, min=0, max=10)
dt = cbind(x1,x2,x3,x4)
#arm reward
arm_1 <-  as.vector(c(-1,9,-8,4))
K1 = crossprod(t(dt),arm_1)
#covariates
dt <-as.data.frame(dt)
#prepare data for regression
M <- dt
M$arm <- 1
M$reward <- K1
#cumpute the model
ker_model <- update_model_for_KernelUCB(choice=1 , M=M ,update_val=100)
#evaluate the model
xnew <- as.data.frame(dt)
ynew <- predict(ker_model$model, xnew)
ytrue <- K1
mean((ynew - K1) ^ 2)  # MSE
plot(density((ynew - K1)))
}
