% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit4abtest.R
\docType{package}
\name{bandit4abtest}
\alias{bandit4abtest}
\title{bandit4abtest Package}
\description{
Contains functions for improving traditional A/B test with methods from
bandit theory.
}
\author{
Emmanuelle Claeys \email{claeys@unistra.fr}
}
