% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit_reward_control.R
\name{bandit_reward_control}
\alias{bandit_reward_control}
\title{Visitor rewards dataframe control}
\usage{
bandit_reward_control(visitor_reward)
}
\arguments{
\item{visitor_reward}{Dataframe of numeric values}
}
\value{
Logical value
}
\description{
Controls data for bandit algorithm. See also
 \code{\link{control_missing_data}} and \code{\link{control_size}}
}
\details{
visitor_reward must be a dataframe without NA values, with a number
 of columns greater than or equal to 2 and a number of rows greater than or
 equal to the number of columns.
}
