% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit_reward_control.R
\name{data_control_context_reward}
\alias{data_control_context_reward}
\title{data_control_context_reward}
\usage{
data_control_context_reward(dt, visitor_reward)
}
\arguments{
\item{dt}{Dataframe of integer numeric or factor values}

\item{visitor_reward}{Dataframe of integer or numeric values}
}
\value{
Logical value
}
\description{
Controls whether the number of rows of dt and visitor_reward
 matches. Returns TRUE if it does, produces an error if not.
}
