% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleRegret.R
\name{simpleRegret}
\alias{simpleRegret}
\title{Return list of regret}
\usage{
simpleRegret(choice, visitorReward)
}
\arguments{
\item{choice}{Integer list}

\item{visitorReward}{dataframe of integer or numeric values}
}
\value{
List of numeric values
}
\description{
Return a list with obtained regret at each iterations
}
\examples{
## Generates 1000 numbers from 2 uniform distributions
set.seed(4434)
K1 <- rnorm(100, 30, .05)
K2 <- rnorm(100, 21, .05)
## Define a dataframe of rewards
visitorReward <- as.data.frame(cbind(K1,K2) )
#Random choices
choice <- sample(c(1,2), 100, replace = TRUE)
simpleRegret(choice=choice,visitorReward=visitorReward)

}
