% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCB.R
\name{UCB}
\alias{UCB}
\title{UCB algorithm}
\usage{
UCB(visitorReward, K = ncol(visitorReward), alpha = 1)
}
\arguments{
\item{visitorReward}{Dataframe of integer or numeric values}

\item{K}{Integer value (optional)}

\item{alpha}{Numeric value (optional)}
}
\value{
\itemize{ List of element:
 \item S:numerical matrix of results ,
 \item choice: choices of UCB,
 \item proba: probability of the chosen arms,
 \item time: time of cumputation,
 \item theta_hat: mean estimated of each arm
 \item theta: real mean of each arm
 }
}
\description{
Control data in visitorReward with \code{\link{bandit_reward_control}}
Stop if something is wrong.
Generate a matrix to save the results (S).
\itemize{ At each iteration
 \item Calculates the arm probabilities
 \item Choose the arm with the maximum upper bound (with \alpha parameter)
 \item Receives a reward in visitorReward for the arm and associated iteration
 \item Updates the results matrix S.
 }
Returns the calculation time.
Review the estimated, actual averages and number of choices for each arm.
See also \code{\link{condition_For_UCB}}, \code{\link{generate_Matrix_SK}},
\code{\link{proba_max_For_UCB}} and \code{\link{play_arm}}.
Require \code{\link{tic}} and \code{\link{toc}} from \code{\link{tictoc}} library
}
\examples{
## Generates 10000 numbers from 2 binomial  distributions
set.seed(4434)
K1 <- rbinom(1000, 1, 0.6)
K2 <- rbinom(1000, 1, 0.7)
## Define a dataframe of rewards
visitorReward <- as.data.frame(cbind(K1,K2) )
#Run UCB algorithm
ucb_alloc  <- UCB(visitorReward,alpha = 10)
ucb_alloc$S
barplot(table(ucb_alloc$choice),main = "Histogram of choices",xlab="arm")
#Upper bound for arm 2 according iterations (red line is the real mean)
plot(x=c(1:length(ucb_alloc$choice[ucb_alloc$choice==2])),
  ucb_alloc$proba[ucb_alloc$choice==2],
  type='l',xlab = 'Time',ylab = 'Upper bound of arm 2')
  lines(c(1:length(ucb_alloc$choice[ucb_alloc$choice==2])),rep(mean(K2),length(ucb_alloc$choice[ucb_alloc$choice==2])),col="red")
ucb_alloc$time
ucb_alloc$theta_hat
ucb_alloc$theta
}
